renames, more scaladoc, ResponseLogger out of its object
