+++ src/main/scala/nl.knaw.dans.easy.deposit/authentication/AuthenticationSupport.scala
@@ -19,7 +19,7 @@ import java.net.URL
 
 import nl.knaw.dans.easy.deposit.authentication.AuthUser.UserState
 import nl.knaw.dans.easy.deposit.authentication.AuthenticationSupport._
-import nl.knaw.dans.easy.deposit.logging.ResponseEnhancedLogging._
+import nl.knaw.dans.easy.deposit.servlets._
 import nl.knaw.dans.lib.error._
 import org.scalatra._
 import org.scalatra.auth.ScentryAuthStore.CookieAuthStore
