+++ src/main/scala/nl.knaw.dans.easy.deposit/logging/PlainCookieHeader.scala
@@ -17,7 +17,7 @@ package nl.knaw.dans.easy.deposit.logging
 
 import org.scalatra.ScalatraBase
 
-trait PlainCookieHeader extends RequestEnhancedLogging {
+trait PlainCookieHeader extends RequestLogger {
   this: ScalatraBase =>
 
   override protected def maskCookieHeader(value: String): String = value
