+++ src/main/scala/nl.knaw.dans.easy.deposit/servlets/AuthServlet.scala
@@ -16,7 +16,6 @@
 package nl.knaw.dans.easy.deposit.servlets
 
 import nl.knaw.dans.easy.deposit.EasyDepositApiApp
-import nl.knaw.dans.easy.deposit.logging.ResponseEnhancedLogging._
 import org.scalatra.NoContent
 
 class AuthServlet(app: EasyDepositApiApp) extends AbstractAuthServlet(app) {
