+++ src/main/scala/nl.knaw.dans.easy.deposit/servlets/ProtectedServlet.scala
@@ -16,7 +16,6 @@
 package nl.knaw.dans.easy.deposit.servlets
 
 import nl.knaw.dans.easy.deposit.EasyDepositApiApp
-import nl.knaw.dans.easy.deposit.logging.ResponseEnhancedLogging._
 import org.scalatra.Unauthorized
 
 class ProtectedServlet(app: EasyDepositApiApp) extends AbstractAuthServlet(app) {
