+++ src/main/scala/nl.knaw.dans.easy.deposit/servlets/UserServlet.scala
@@ -18,7 +18,6 @@ package nl.knaw.dans.easy.deposit.servlets
 import nl.knaw.dans.easy.deposit.EasyDepositApiApp
 import nl.knaw.dans.easy.deposit.docs.JsonUtil._
 import nl.knaw.dans.easy.deposit.docs.UserInfo
-import nl.knaw.dans.easy.deposit.logging.ResponseEnhancedLogging._
 import nl.knaw.dans.lib.error._
 import org.scalatra._
