+++ src/main/scala/nl.knaw.dans.easy.deposit/servlets/package.scala
@@ -20,9 +20,10 @@ import java.nio.file.Files
 import java.util.zip.{ ZipEntry, ZipException, ZipInputStream }
 
 import better.files.File
+import javax.servlet.http.{ HttpServletRequest, HttpServletResponse }
 import nl.knaw.dans.easy.deposit.docs.JsonUtil.InvalidDocumentException
+import nl.knaw.dans.easy.deposit.logging.ResponseLogger
 import nl.knaw.dans.easy.deposit.servlets.DepositServlet.{ BadRequestException, ZipMustBeOnlyFileException }
-import nl.knaw.dans.lib.logging.DebugEnhancedLogging
 import org.scalatra.servlet.FileItem
 import org.scalatra.util.RicherString._
 import org.scalatra.{ ActionResult, BadRequest, InternalServerError }
@@ -30,7 +31,7 @@ import resource.{ ManagedResource, managed }
 
 import scala.util.{ Failure, Success, Try }
 
-package object servlets extends DebugEnhancedLogging {
+package object servlets extends ResponseLogger {
 
   private val extensionZipPattern = ".*.g?z(ip)?"
   private val pre = "(x-)?g?zip"
@@ -133,6 +134,13 @@ package object servlets extends DebugEnhancedLogging {
     private def skipLeadingEmptyFormFields(): Unit = {
       // forward pointer after check
       while (fileItems.headOption.exists(_.name.isBlank)) { fileItems.next() }
+    }
+  }
+
+  implicit class RichActionResult(val actionResult: ActionResult) extends AnyVal {
+
+    def logResponse(implicit request: HttpServletRequest, response: HttpServletResponse): ActionResult = {
+      logActionResult(actionResult)
     }
   }
 }
